+++ README/CHANGELOG.md
@@ -100,4 +100,24 @@ Initial Release
 - API updated to v3.1-1
 - a lot of typo fixes
 - removed Security.md, because it was useless, and I don't really remember why I even added it
-- added all files to the release page (thanks to Egsagon for telling me that I should do that :) 
+- added all files to the release page (thanks to Egsagon for telling me that I should do that :)
+
+# 2.3 
+
+- fixed some issues...
+
+
+# 2.4
+
+- If you use the file / model - user - channel functionality, then the TreeWidget will be used
+  to let you select the videos that you want to download instead of downloading everything
+
+- fixed an issue in the termux build script
+- API updated to v3.1-4
+- You can now select if you want to have a delay or not (enabling it is recommended!)
+- OS error is fixed (FOR REAL!)
+- Sentry strips out sensitive information and now only the exception, lines of code, server name is reported
+  (Although I need to still test that.)
+
+The next update will focus more on features / compatibility to other systems.
+I hope that most issues are now fixed.